requirements.txt
flask>=3.0.0
requests>=2.31.0
python-nmap>=0.7.1
waitress>=3.0.0
pywebview>=5.1
